Expand utils.lua.
-   Expand inline docs.
-   Add setCirclePos and vectorFromAngle.
-   Call utils.lua from scripts that use setCirclePos and vectorFronAngle.
