+++ scripts/utils.lua
@@ -3,47 +3,123 @@
 
 --[[ These functions should be as generic as possible, so they are highly usable. --]]
 
---Distance function. Can be used 4 ways:
---distance(obj1, obj2)
---distance(obj, x, y)
---distance(x, y, obj)
---distance(x1, y1, x2, y2)
+-- Given enough information, find the distance between two positions.
+-- This function can be called four ways:
+--
+-- distance(obj1, obj2)
+--   Returns the distance between two objects.
+--
+--   obj1, obj2: Two objects. Calls getPosition() on each.
+--
+--   Example:
+--     rock1 = Asteroid():setPosition(-100, 100)
+--     rock2 = Asteroid():setPosition(0, 100)
+--     distance(rock1, rock2) -- Returns 100
+--
+-- distance(obj, x, y)
+-- distance(x, y, obj)
+--   Find the distance from an object to a position,
+--   or vice versa.
+--
+--   obj: An object. Calls getPosition() on it.
+--   x, y: Coordinates of a position.
+--
+--   Example:
+--     distance(rock1, 0, 100) -- Returns 100
+--     distance(0, 100, rock1) -- Returns 100
+--
+-- distance(x1, y1, x2, y2)
+--   Find the distance between two positions.
+--
+--   x1, y1: Origin position's coordinates.
+--   x2, y2: Destination position's coordinates.
+--
+--   Example:
+--     distance(-100, 100, 0, 100) -- Returns 100
 function distance(a, b, c, d)
     local x1, y1 = 0, 0
     local x2, y2 = 0, 0
     if type(a) == "table" and type(b) == "table" then
-        --both a and b are tables.
-        --Assume distance(obj1, obj2)
+        -- a and b are bth tables.
+        -- Assume distance(obj1, obj2)
         x1, y1 = a:getPosition()
         x2, y2 = b:getPosition()
     elseif type(a) == "table" and type(b) == "number" and type(c) == "number" then
-        --Assume distance(obj, x, y)
+        -- Assume distance(obj, x, y)
         x1, y1 = a:getPosition()
         x2, y2 = b, c
     elseif type(a) == "number" and type(b) == "number" and type(c) == "table" then
-        --Assume distance(x, y, obj)
+        -- Assume distance(x, y, obj)
         x1, y1 = a, b:getPosition()
         x2, y2 = c:getPosition()
     elseif type(a) == "number" and type(b) == "number" and type(c) == "number" and type(d) == "number" then
-        --both a and b are tables.
-        --Assume distance(obj1, obj2)
+        -- a and b are both tables.
+        -- Assume distance(obj1, obj2)
         x1, y1 = a, b
         x2, y2 = c, d
     else
-        -- Not a valid usage of the distance function, throw an error.
+        -- Not a valid use of the distance function. Throw an error.
         print(type(a), type(b), type(c), type(d))
-        error("distance function incorrectly used", 2)
+        error("distance() function used incorrectly", 2)
     end
     local xd, yd = (x1 - x2), (y1 - y2)
     return math.sqrt(xd * xd + yd * yd)
 end
 
---Function to create objects in a line formation.
--- Per default this creates a grid/line formation of objects. For example a mine field will look like:
--- createObjectsOnLine(0, 0, 10000, 0, 1000, Mine, 4)
--- As this creates 4 rows of mines from 0,0 to 10000,0 with a spacing of 1000 between each mine.
--- The randomize parameter can be used to add more chaos to the pattern. This works well for asteroids, example:
--- createObjectsOnLine(0, 0, 10000, 0, 300, Asteroid, 4, 100, 800)
+-- Given an angle and length, return a relative vector (x, y coordinates).
+--
+-- vectorFromAngle(angle, length)
+--   angle: Relative heading, in degrees
+--   length: Relative distance, in thousandths of an in-game unit (1000 = 1U)
+--
+-- Example: For relative x and y coordinates 1000 units away at a heading of
+-- 45 degrees, run:
+--   vectorFromAngle(45, 1000).
+function vectorFromAngle(angle, length)
+    return math.cos(angle / 180 * math.pi) * length, math.sin(angle / 180 * math.pi) * length
+end
+
+-- Place an object relative to a vector. Returns the object with its position
+-- set to the resulting coordinates.
+--
+-- setCirclePos(obj, x, y, angle, distance)
+--   obj: An object.
+--   x, y: Origin coordinates.
+--   angle, distance: Relative heading and distance from the origin.
+--
+-- Returns the object with its position set to the resulting coordinates, by
+-- calling setPosition().
+--
+-- Example: To create a space station 10000 units from coordinates 100, -100
+-- at a heading of 45 degrees, run
+--   setCirclePos(SpaceStation():setTemplate("Small Station"):setFaction("Independent"), 100, -100, 45, 10000)
+function setCirclePos(obj, x, y, angle, distance)
+    dx, dy = vectorFromAngle(angle, distance)
+    return obj:setPosition(x + dx, y + dy)
+end
+
+-- Create objects along a line between two vectors, optionally with grid
+-- placement and randomization.
+--
+-- createObjectsOnLine(x1, y1, x2, y2, spacing, object_type, rows, chance, randomize)
+--   x1, y1: Starting coordinates
+--   x2, y2: Ending coordinates
+--   spacing: The distance between each object.
+--   object_type: The object type. Calls `object_type():setPosition()`.
+--   rows (optional): The number of rows, minimum 1. Defaults to 1.
+--   chance (optional): The percentile chance an object will be created,
+--     minimum 1. Defaults to 100 (always).
+--   randomize (optional): If present, randomize object placement by this
+--     amount. Defaults to 0 (grid).
+--
+--   Examples: To create a mine field, run:
+--     createObjectsOnLine(0, 0, 10000, 0, 1000, Mine, 4)
+--   This creates 4 rows of mines from 0,0 to 10000,0, with mines spaced 1U
+--   apart.
+--
+--   The `randomize` parameter adds chaos to the pattern. This works well for
+--   asteroid fields:
+--     createObjectsOnLine(0, 0, 10000, 0, 300, Asteroid, 4, 100, 800)
 function createObjectsOnLine(x1, y1, x2, y2, spacing, object_type, rows, chance, randomize)
     if rows == nil then rows = 1 end
     if chance == nil then chance = 100 end
@@ -62,8 +138,11 @@ function createObjectsOnLine(x1, y1, x2, y2, spacing, object_type, rows, chance,
     end
 end
 
---Function to merge data from table B into table A. Every key that is not set in table A is filled by the data from table B.
--- this function can be used to fill in incomplete configuration data. See the comms_station script as example.
+-- Merge data from table B into table A. Every key not set in table A is filled
+-- by the data from table B.
+--
+-- This function can fill in incomplete configuration data. See the
+-- comms_station.lua script as example.
 function mergeTables(table_a, table_b)
     for key, value in pairs(table_b) do
         if table_a[key] == nil then
